2748 - 피보나치 수 2 (반환형 주의)

+++ algorithm/2748.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 long long arr[90];
 
-int fibo(int n) {
+long long fibo(int n) {
 	if (n == 1 || n == 2)
 		return 1;
 	if (arr[n] != 0)
